Fix a leak in Lobsters benchmark
Profiling the lobsters benchmark I'm seeing a fairly large time spent
joining tags in TaggedLogger.

That intruigued me, and by digging I noticed the requests UUIDs kept
being appended but never removed.

So for every single request we have one more tag, this is skewing
the benchmark a bit, as we spend much more time than we should in
logging.