Update lobsters
THe initial snapshot was from a few years back, the application changed
a lot since then.

I had to run a bunch of migrations and had to fiddle with various sqlite3/msyql
incompatibilities. Also some route changes, but with a bit of effort I managed
to make it work again.

What is interesting is it seems lobste.rs became 2-3-x as slow in these
couple years, I haven't yet profiled to know why.

This branch:

```
itr:   time
 #1: 1532ms
 #2: 1013ms
 #3:  910ms
 #4:  852ms
 #5:  851ms
 #6:  901ms
 #7:  857ms
 #8:  886ms
 #9:  875ms
```

main:

```
itr:   time
 #1:  884ms
 #2:  482ms
 #3:  344ms
 #4:  358ms
 #5:  300ms
 #6:  316ms
 #7:  315ms
 #8:  328ms
 #9:  309ms
```