Integrate debug counters
If we compile ruby with debug counters and use that ruby for running yjit-bench,
there's a mess of output due to the ruby subprocesses that get run (ex: `bundle install`).
This is because by default, the output is shown at ruby exit. This change adds a new option
`--show-debug-counters` that disables all debug counter output for everything except the
benchmark scripts. The counters get reset before each benchmark script is run so that the
counters during ruby initialization and prelude are not taken into account.