Add benchmarks/ractor directory with 2 JSON benchmarks
Add JSON benchmarks:
  parse_float
  parse_string

Add to ractor harness, make it possible to use gemfiles in different
directories and make it easy to do so. This way we don't have to create
directories for each benchmark.

Improve run_benchmarks.rb, allow it to take a regular expression for --name_filters option
Ex:
$ WARMUP_ITRS=10 ruby run_benchmarks.rb --category=ractor-only --name_filters=/json/

Allow it to take --skip-yjit option  to only run with interpreter.